[refactor] Move navigation package to another place
