+++ app/src/main/java/com/makentoshe/habrachan/di/AppActivityModule.kt
@@ -2,7 +2,7 @@ package com.makentoshe.habrachan.di
 
 import com.makentoshe.habrachan.AppActivity
 import com.makentoshe.habrachan.R
-import com.makentoshe.habrachan.common.model.navigation.Navigator
+import com.makentoshe.habrachan.common.navigation.Navigator
 import toothpick.config.Module
 import toothpick.ktp.binding.bind
