+++ app/src/main/java/com/makentoshe/habrachan/model/MainFlowScreen.kt
@@ -1,9 +1,8 @@
 package com.makentoshe.habrachan.model
 
 import androidx.fragment.app.Fragment
-import com.makentoshe.habrachan.common.model.navigation.Screen
+import com.makentoshe.habrachan.common.navigation.Screen
 import com.makentoshe.habrachan.view.MainFlowFragment
-import ru.terrakok.cicerone.android.support.SupportAppScreen
 
 class MainFlowScreen : Screen() {
     override val fragment: Fragment
