+++ app/src/main/java/com/makentoshe/habrachan/model/account/AccountFragmentScreen.kt
@@ -1,7 +1,7 @@
 package com.makentoshe.habrachan.model.account
 
 import androidx.fragment.app.Fragment
-import com.makentoshe.habrachan.common.model.navigation.Screen
+import com.makentoshe.habrachan.common.navigation.Screen
 import com.makentoshe.habrachan.view.account.AccountFragment
 
 class AccountScreen : Screen() {
