+++ app/src/main/java/com/makentoshe/habrachan/model/menu/MenuScreen.kt
@@ -1,7 +1,7 @@
 package com.makentoshe.habrachan.model.menu
 
 import androidx.fragment.app.Fragment
-import com.makentoshe.habrachan.common.model.navigation.Screen
+import com.makentoshe.habrachan.common.navigation.Screen
 import com.makentoshe.habrachan.view.menu.MenuFragment
 
 class MenuScreen : Screen() {
