+++ app/src/main/java/com/makentoshe/habrachan/model/posts/PostsScreen.kt
@@ -1,7 +1,7 @@
 package com.makentoshe.habrachan.model.posts
 
 import androidx.fragment.app.Fragment
-import com.makentoshe.habrachan.common.model.navigation.Screen
+import com.makentoshe.habrachan.common.navigation.Screen
 import com.makentoshe.habrachan.view.posts.PostsFragment
 
 class PostsScreen(private val page: Int) : Screen(){
